+++ runtime/onert/api/src/nnfw_debug.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "nnfw_api_internal.h"
+#include "Config.h"
 
 #include <util/ConfigSource.h>
 
@@ -28,11 +28,11 @@
 NNFW_STATUS nnfw_set_config(nnfw_session *session, const char *key, const char *value)
 {
   NNFW_RETURN_ERROR_IF_NULL(session);
-  return session->set_config(key, value);
+  return onert::api::Config(session).setConfig(key, value);
 }
 
 NNFW_STATUS nnfw_get_config(nnfw_session *session, const char *key, char *value, size_t value_size)
 {
   NNFW_RETURN_ERROR_IF_NULL(session);
-  return session->get_config(key, value, value_size);
+  return onert::api::Config(session).getConfig(key, value, value_size);
 }
